resume-viewer/requirements.txt
setuptools==58.1.0
wheel==0.44.0
toml==0.10.2
python-dotenv==1.0.1
fastapi==0.115.0
python-multipart==0.0.9
uvicorn==0.30.6
pymupdf==1.24.10
python-docx==1.1.2
langchain_core==0.3.2
langchain_openai==0.2.1
mysql-connector-python==9.0.0
boto3==1.35.25
